git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/branches/branch-1.2@1719 7a7537e8-13f0-0310-91df-b6672ffda945

+++ pdftops/PDFDoc.h
@@ -61,8 +61,8 @@ public:
   int getNumPages() { return catalog->getNumPages(); }
 
   // Display a page.
-  void displayPage(OutputDev *out, int page, int zoom, int rotate,
-		   GBool doLinks);
+  void displayPage(OutputDev *out, int page, double zoom,
+		   int rotate, GBool doLinks);
 
   // Display a range of pages.
   void displayPages(OutputDev *out, int firstPage, int lastPage,
